google-cloud-functions/verify-otp/requirements.txt
functions-framework==3.5.0
flask==3.0.0
google-cloud-firestore==2.14.0
